+++ docs/conf.py
@@ -62,7 +62,6 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 autodoc_mock_imports = [
     "openff.models",
     "rdkit",
-    "networkx",
 ]
 
 
@@ -71,10 +70,10 @@ autodoc_mock_imports = [
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 #
-html_theme = "ofe_sphinx_theme"
-html_theme_options = {
-    "accent_color": "FeelingFabulous",
-}
+html_theme = "furo"
+# html_theme_options = {
+#    "accent_color": "FeelingFabulous",
+# }
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,