+++ gufe/mapping/errors.py
@@ -1,8 +1,6 @@
 # This code is part of OpenFE and is licensed under the MIT license.
 # For details, see https://github.com/OpenFreeEnergy/gufe
 
-from typing import TypeAlias
 
-from rdkit import Chem
-
-RDKitMol: TypeAlias = Chem.rdchem.Mol
+class AtomMappingError(Exception):
+    """Error when an atom mapping could not be created."""