+++ gufe/protocols/protocol.py
@@ -92,9 +92,15 @@ class Protocol(GufeTokenizable):
     - `_create`
     - `_gather`
     - `_default_settings`
+
+    Additionally, the `_settings_cls` must be set to the intended
+    subclass of `SettingsBaseModel`. This attribute is validated
+    during the instantiation of the Protocol.
+
     """
 
     _settings: Settings
+    _settings_cls: type[SettingsBaseModel]
     result_cls: type[ProtocolResult]
     """Corresponding `ProtocolResult` subclass."""
 
@@ -111,6 +117,17 @@ class Protocol(GufeTokenizable):
         Once the Protocol object is created, the input Settings are frozen,
         so should be finalised before creating the Protocol instance.
         """
+
+        if not hasattr(self.__class__, "_settings_cls"):
+            raise NotImplementedError(
+                f"class `{self.__class__.__qualname__}` must implement the `_settings_cls` attribute."
+            )
+
+        if not isinstance(settings, self._settings_cls):
+            raise ValueError(
+                f"`{self.__class__.__qualname__}` expected a `{self._settings_cls.__qualname__}` instance."
+            )
+
         self._settings = settings.frozen_copy()
 
     @property
@@ -155,8 +172,8 @@ class Protocol(GufeTokenizable):
         self,
         stateA: ChemicalSystem,
         stateB: ChemicalSystem,
-        mapping: Optional[Union[ComponentMapping, list[ComponentMapping]]],
-        extends: Optional[ProtocolDAGResult] = None,
+        mapping: ComponentMapping | list[ComponentMapping] | None,
+        extends: ProtocolDAGResult | None = None,
     ) -> list[ProtocolUnit]:
         """Method to override in custom :class:`Protocol` subclasses.
 
@@ -191,10 +208,10 @@ class Protocol(GufeTokenizable):
         *,
         stateA: ChemicalSystem,
         stateB: ChemicalSystem,
-        mapping: Optional[Union[ComponentMapping, list[ComponentMapping], dict[str, ComponentMapping]]],
-        extends: Optional[ProtocolDAGResult] = None,
-        name: Optional[str] = None,
-        transformation_key: Optional[GufeKey] = None,
+        mapping: ComponentMapping | list[ComponentMapping] | dict[str, ComponentMapping] | None,
+        extends: ProtocolDAGResult | None = None,
+        name: str | None = None,
+        transformation_key: GufeKey | None = None,
     ) -> ProtocolDAG:
         r"""Prepare a `ProtocolDAG` with all information required for execution.