+++ gufe/protocols/protocoldag.py
@@ -14,22 +14,23 @@ from typing import Any, Optional, Union
 import networkx as nx
 
 from ..tokenization import GufeKey, GufeTokenizable
+from .errors import MissingUnitResultError, ProtocolUnitFailureError
 from .protocolunit import Context, ProtocolUnit, ProtocolUnitFailure, ProtocolUnitResult
 
 
 class DAGMixin:
     _protocol_units: list[ProtocolUnit]
 
-    _name: Optional[str]
+    _name: str | None
     _graph: nx.DiGraph
 
     # labels for identifying source of this DAG
 
     ## key of the Transformation that this DAG corresponds to
-    _transformation_key: Union[GufeKey, None]
+    _transformation_key: GufeKey | None
 
     ## key of the ProtocolDAG this DAG extends
-    _extends_key: Optional[GufeKey]
+    _extends_key: GufeKey | None
 
     @staticmethod
     def _build_graph(nodes):
@@ -48,7 +49,7 @@ class DAGMixin:
         return reversed(list(nx.lexicographical_topological_sort(graph, key=lambda pu: pu.key)))
 
     @property
-    def name(self) -> Optional[str]:
+    def name(self) -> str | None:
         """Optional identifier."""
         return self._name
 
@@ -68,7 +69,7 @@ class DAGMixin:
         return list(self._iterate_dag_order(self._graph))
 
     @property
-    def transformation_key(self) -> Union[GufeKey, None]:
+    def transformation_key(self) -> GufeKey | None:
         """
         The `GufeKey` of the `Transformation` this object performs.
 
@@ -81,7 +82,7 @@ class DAGMixin:
         return self._transformation_key
 
     @property
-    def extends_key(self) -> Union[GufeKey, None]:
+    def extends_key(self) -> GufeKey | None:
         """The `GufeKey` of the `ProtocolDAGResult` this object extends.
 
         If `None`, then this object does not extend from a result at all.
@@ -111,9 +112,9 @@ class ProtocolDAGResult(GufeTokenizable, DAGMixin):
         *,
         protocol_units: list[ProtocolUnit],
         protocol_unit_results: list[ProtocolUnitResult],
-        transformation_key: Union[GufeKey, None],
-        extends_key: Optional[GufeKey] = None,
-        name: Optional[str] = None,
+        transformation_key: GufeKey | None,
+        extends_key: GufeKey | None = None,
+        name: str | None = None,
     ):
         self._name = name
         self._protocol_units = protocol_units
@@ -211,22 +212,24 @@ class ProtocolDAGResult(GufeTokenizable, DAGMixin):
 
         Raises
         ------
-        KeyError
-          if either there are no results, or only failures
+        MissingUnitResultError:
+          if there are no results for that protocol unit
+        ProtocolUnitFailureError:
+          if there are only failures for that protocol unit
         """
         try:
             units = self._unit_result_mapping[protocol_unit]
         except KeyError:
-            raise KeyError("No such `protocol_unit` present")
+            raise MissingUnitResultError(f"No such `protocol_unit`:{protocol_unit} present")
         else:
             for u in units:
                 if u.ok():
                     return u
             else:
-                raise KeyError("No success for `protocol_unit` found")
+                raise ProtocolUnitFailureError(f"No success for `protocol_unit`:{protocol_unit} found")
 
     def unit_to_all_results(self, protocol_unit: ProtocolUnit) -> list[ProtocolUnitResult]:
-        """Return all results (sucess and failure) for a given Unit.
+        """Return all results (success and failure) for a given Unit.
 
         Returns
         -------
@@ -235,19 +238,19 @@ class ProtocolDAGResult(GufeTokenizable, DAGMixin):
 
         Raises
         ------
-        KeyError
+        MissingUnitResultError
           if no results present for a given unit
         """
         try:
             return self._unit_result_mapping[protocol_unit]
         except KeyError:
-            raise KeyError("No such `protocol_unit` present")
+            raise MissingUnitResultError(f"No such `protocol_unit`:{protocol_unit} present")
 
     def result_to_unit(self, protocol_unit_result: ProtocolUnitResult) -> ProtocolUnit:
         try:
             return self._result_unit_mapping[protocol_unit_result]
         except KeyError:
-            raise KeyError("No such `protocol_unit_result` present")
+            raise MissingUnitResultError(f"No such `protocol_unit_result`:{protocol_unit_result} present")
 
     def ok(self) -> bool:
         # ensure that for every protocol unit, there is an OK result object
@@ -293,9 +296,9 @@ class ProtocolDAG(GufeTokenizable, DAGMixin):
         self,
         *,
         protocol_units: list[ProtocolUnit],
-        transformation_key: Union[GufeKey, None],
-        extends_key: Optional[GufeKey] = None,
-        name: Optional[str] = None,
+        transformation_key: GufeKey | None,
+        extends_key: GufeKey | None = None,
+        name: str | None = None,
     ):
         """Create a new `ProtocolDAG`
 
@@ -440,7 +443,7 @@ def execute_DAG(
 
 
 def _pu_to_pur(
-    inputs: Union[dict[str, Any], list[Any], ProtocolUnit],
+    inputs: dict[str, Any] | list[Any] | ProtocolUnit,
     mapping: dict[GufeKey, ProtocolUnitResult],
 ):
     """Convert each `ProtocolUnit` found within `inputs` to its corresponding