+++ gufe/tests/test_smallmoleculecomponent.py
@@ -12,6 +12,7 @@ except ImportError:
 else:
     HAS_OFFTK = True
 import json
+import logging
 import os
 from unittest import mock
 
@@ -24,6 +25,7 @@ from gufe import SmallMoleculeComponent
 from gufe.components.explicitmoleculecomponent import _ensure_ofe_name
 from gufe.tokenization import TOKENIZABLE_REGISTRY
 
+from .test_explicitmoleculecomponent import ExplicitMoleculeComponentMixin
 from .test_tokenization import GufeTokenizableTestsMixin
 
 
@@ -74,7 +76,7 @@ def test_ensure_ofe_name(internal, rdkit_name, name, expected, recwarn):
     assert rdkit.GetProp("ofe-name") == out_name
 
 
-class TestSmallMoleculeComponent(GufeTokenizableTestsMixin):
+class TestSmallMoleculeComponent(GufeTokenizableTestsMixin, ExplicitMoleculeComponentMixin):
 
     cls = SmallMoleculeComponent
     repr = "SmallMoleculeComponent(name=ethane)"
@@ -241,15 +243,16 @@ class TestSmallMoleculeComponentConversion:
 @pytest.mark.skipif(not HAS_OFFTK, reason="no openff tookit available")
 class TestSmallMoleculeComponentPartialCharges:
     @pytest.fixture(scope="function")
-    def charged_off_ethane(self, ethane):
-        off_ethane = ethane.to_openff()
+    def charged_off_ethane(self, named_ethane):
+        off_ethane = named_ethane.to_openff()
         off_ethane.assign_partial_charges(partial_charge_method="am1bcc")
         return off_ethane
 
-    def test_partial_charges_warning(self, charged_off_ethane):
-        matchmsg = "Partial charges have been provided"
-        with pytest.warns(UserWarning, match=matchmsg):
-            SmallMoleculeComponent.from_openff(charged_off_ethane)
+    def test_partial_charges_logging(self, charged_off_ethane, caplog):
+        caplog.set_level(logging.INFO)
+        SmallMoleculeComponent.from_openff(charged_off_ethane)
+
+        assert "Partial charges are present for SmallMoleculeComponent-" in caplog.text
 
     def test_partial_charges_zero_warning(self, charged_off_ethane):
         charged_off_ethane.partial_charges[:] = 0 * unit.elementary_charge
@@ -257,8 +260,9 @@ class TestSmallMoleculeComponentPartialCharges:
         with pytest.warns(UserWarning, match=matchmsg):
             SmallMoleculeComponent.from_openff(charged_off_ethane)
 
-    def test_partial_charges_not_formal_error(self, charged_off_ethane):
-        charged_off_ethane.partial_charges[:] = 1 * unit.elementary_charge
+    @pytest.mark.parametrize("wrong_charge_val", [-1, 1])
+    def test_partial_charges_not_formal_error(self, charged_off_ethane, wrong_charge_val):
+        charged_off_ethane.partial_charges[:] = wrong_charge_val * unit.elementary_charge
         with pytest.raises(ValueError, match="Sum of partial charges"):
             SmallMoleculeComponent.from_openff(charged_off_ethane)
 
@@ -270,7 +274,7 @@ class TestSmallMoleculeComponentPartialCharges:
         with pytest.raises(ValueError, match="Incorrect number of"):
             SmallMoleculeComponent.from_rdkit(mol)
 
-    def test_partial_charges_applied_to_atoms(self):
+    def test_partial_charges_applied_to_atoms(self, caplog):
         """
         Make sure that charges set at the molecule level
         are transferred to atoms and picked up by openFF.
@@ -279,9 +283,11 @@ class TestSmallMoleculeComponentPartialCharges:
         Chem.AllChem.Compute2DCoords(mol)
         # add some fake charges at the molecule level
         mol.SetProp("atom.dprop.PartialCharge", "-1 0.25 0.25 0.25 0.25")
-        matchmsg = "Partial charges have been provided"
-        with pytest.warns(UserWarning, match=matchmsg):
-            ofe = SmallMoleculeComponent.from_rdkit(mol)
+        caplog.set_level(logging.INFO)
+
+        ofe = SmallMoleculeComponent.from_rdkit(mol)
+        assert "Partial charges are present for SmallMoleculeComponent-" in caplog.text
+
         # convert to openff and make sure the charges are set
         off_mol = ofe.to_openff()
         assert off_mol.partial_charges is not None
@@ -338,6 +344,36 @@ class TestSmallMoleculeSerialization:
         d = phenol.to_dict()
 
         assert isinstance(d, dict)
+
+    def test_to_dict_hybridization(self, phenol):
+        """
+        Make sure dict round trip saves the hybridization
+        <https://github.com/OpenFreeEnergy/gufe/issues/407>
+        """
+        phenol_dict = phenol.to_dict()
+        TOKENIZABLE_REGISTRY.clear()
+        new_phenol = SmallMoleculeComponent.from_dict(phenol_dict)
+        for atom in new_phenol.to_rdkit().GetAtoms():
+            if atom.GetAtomicNum() == 6:
+                assert atom.GetHybridization() == Chem.rdchem.HybridizationType.SP2
+
+    def test_from_dict_missing_hybridization(self, phenol):
+        """
+        For backwards compatibility make sure we can create an SMC with missing hybridization info.
+        """
+        phenol_dict = phenol.to_dict()
+        new_atoms = []
+        for atom in phenol_dict["atoms"]:
+            # remove the hybridization atomic info which should be at index 7
+            new_atoms.append(tuple([atom_info for i, atom_info in enumerate(atom) if i != 7]))
+        phenol_dict["atoms"] = new_atoms
+        with pytest.warns(match="The atom hybridization data was not found and has been set to unspecified."):
+            new_phenol = SmallMoleculeComponent.from_dict(phenol_dict)
+        # they should be different objects due to the missing hybridization info
+        assert new_phenol != phenol
+        # make sure the rdkit objects are different
+        for atom_hybrid, atom_no_hybrid in zip(phenol.to_rdkit().GetAtoms(), new_phenol.to_rdkit().GetAtoms()):
+            assert atom_hybrid.GetHybridization() != atom_no_hybrid.GetHybridization()
 
     @pytest.mark.skipif(not HAS_OFFTK, reason="no openff toolkit available")
     def test_deserialize_roundtrip(self, toluene, phenol):