+++ gufe/tests/test_tokenization.py
@@ -92,7 +92,7 @@ class GufeTokenizableTestsMixin(abc.ABC):
 
     # set this to the `GufeTokenizable` subclass you are testing
     cls: type[GufeTokenizable]
-    repr: Optional[str]
+    repr: str | None
 
     @pytest.fixture
     def instance(self):
@@ -238,7 +238,6 @@ class TestGufeTokenizable(GufeTokenizableTestsMixin):
         patch_loc = "gufe.tokenization.TOKENIZABLE_REGISTRY"
         registry = dict(TOKENIZABLE_REGISTRY)
         with mock.patch.dict(patch_loc, registry, clear=True):
-            # import pdb; pdb.set_trace()
             leaf._set_key("qux")
             assert leaf.key == "qux"
             assert TOKENIZABLE_REGISTRY["qux"] is leaf
@@ -284,6 +283,14 @@ class TestGufeTokenizable(GufeTokenizableTestsMixin):
         assert recreated == self.cont
         assert recreated is self.cont
 
+    def test_from_json_string_dict(self):
+        """Test that we can still load json-serialized dict representations."""
+        with pytest.warns(UserWarning, match="keyed-chain deserialization failed"):
+            recreated = self.cls.from_json(content=json.dumps(self.expected_deep, cls=JSON_HANDLER.encoder))
+
+        assert recreated == self.cont
+        assert recreated is self.cont
+
     def test_to_json_file(self, tmpdir):
         file_path = tmpdir / "container.json"
         self.cont.to_json(file=file_path)
@@ -300,6 +307,20 @@ class TestGufeTokenizable(GufeTokenizableTestsMixin):
             cls=JSON_HANDLER.encoder,
         )
         recreated = self.cls.from_json(file=file_path)
+
+        assert recreated == self.cont
+        assert recreated is self.cont
+
+    def test_from_json_file_dict(self, tmpdir):
+        """Test that we can still load json-serialized dict representations from files."""
+        file_path = tmpdir / "container.json"
+        json.dump(
+            self.expected_deep,
+            file_path.open(mode="w"),
+            cls=JSON_HANDLER.encoder,
+        )
+        with pytest.warns(UserWarning, match="keyed-chain deserialization failed"):
+            recreated = self.cls.from_json(file=file_path)
 
         assert recreated == self.cont
         assert recreated is self.cont