+++ gufe/transformations/transformation.py
@@ -1,6 +1,7 @@
 # This code is part of OpenFE and is licensed under the MIT license.
 # For details, see https://github.com/OpenFreeEnergy/gufe
 
+import abc
 import json
 import warnings
 from collections.abc import Iterable
@@ -13,61 +14,191 @@ from ..tokenization import JSON_HANDLER, GufeTokenizable
 from ..utils import ensure_filelike
 
 
-class Transformation(GufeTokenizable):
+class TransformationBase(GufeTokenizable):
+    _protocol: Protocol
+    _name: str | None
+
+    def __init__(
+        self,
+        protocol: Protocol,
+        name: str | None = None,
+    ):
+        """Transformation base class.
+
+        Parameters
+        ----------
+        protocol : Protocol
+            The sampling method to use for the transformation.
+        name : str, optional
+            A human-readable name for this transformation.
+
+        """
+        self._protocol = protocol
+        self._name = name
+
+    @classmethod
+    def _defaults(cls):
+        return super()._defaults()
+
+    @property
+    def name(self) -> str | None:
+        """Optional identifier for the transformation; used as part of its hash.
+
+        Set this to a unique value if adding multiple, otherwise identical
+        transformations to the same :class:`.AlchemicalNetwork` to avoid
+        deduplication.
+        """
+        return self._name
+
+    @classmethod
+    def _from_dict(cls, d: dict):
+        return cls(**d)
+
+    @property
+    @abc.abstractmethod
+    def stateA(self) -> ChemicalSystem:
+        """The starting :class:`.ChemicalSystem` for the transformation."""
+        raise NotImplementedError
+
+    @property
+    @abc.abstractmethod
+    def stateB(self) -> ChemicalSystem:
+        """The ending :class:`.ChemicalSystem` for the transformation."""
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def create(
+        self,
+        *,
+        extends: ProtocolDAGResult | None = None,
+        name: str | None = None,
+    ) -> ProtocolDAG:
+        """
+        Returns a :class:`.ProtocolDAG` executing this ``Transformation.protocol``.
+        """
+        raise NotImplementedError
+
+    @property
+    def protocol(self) -> Protocol:
+        """The :class:`.Protocol` used to perform the transformation.
+
+        The protocol estimates the free energy differences between ``stateA``
+        and ``stateB`` :class:`.ChemicalSystem` objects. It includes all
+        details needed to perform required simulations/calculations and encodes
+        the alchemical or non-alchemical pathway used.
+
+        """
+        return self._protocol
+
+    def gather(self, protocol_dag_results: Iterable[ProtocolDAGResult]) -> ProtocolResult:
+        r"""Gather multiple :class:`.ProtocolDAGResult` \s into a single
+        :class:`.ProtocolResult`.
+
+        Parameters
+        ----------
+        protocol_dag_results : Iterable[ProtocolDAGResult]
+            The :class:`.ProtocolDAGResult` objects to assemble aggregate
+            quantities from.
+
+        Returns
+        -------
+        ProtocolResult
+            Aggregated results from many :class:`.ProtocolDAGResult` objects,
+            all from a given :class:`.Protocol`.
+
+        """
+        return self.protocol.gather(protocol_dag_results=protocol_dag_results)
+
+    def dump(self, file):
+        """Dump this Transformation to a JSON file.
+
+        Note that this is not space-efficient: for example, any ``Component``
+        which is used in both ``ChemicalSystem`` objects will be represented
+        twice in the JSON output.
+
+        Parameters
+        ----------
+        file : Union[PathLike, FileLike]
+            A pathlike of filelike to save this transformation to.
+        """
+        warnings.warn(
+            ("use of this method is deprecated; instead use `to_json`"),
+            DeprecationWarning,
+        )
+        with ensure_filelike(file, mode="w") as f:
+            json.dump(self.to_dict(), f, cls=JSON_HANDLER.encoder, sort_keys=True)
+
+    @classmethod
+    def load(cls, file):
+        """Create a Transformation from a JSON file.
+
+        Parameters
+        ----------
+        file : Union[PathLike, FileLike]
+            A pathlike or filelike to read this transformation from.
+        """
+        warnings.warn(
+            ("use of this method is deprecated; instead use `from_json`"),
+            DeprecationWarning,
+        )
+        with ensure_filelike(file, mode="r") as f:
+            dct = json.load(f, cls=JSON_HANDLER.decoder)
+
+        return cls.from_dict(dct)
+
+
+class Transformation(TransformationBase):
     _stateA: ChemicalSystem
     _stateB: ChemicalSystem
-    _name: Optional[str]
-    _mapping: Optional[Union[ComponentMapping, list[ComponentMapping]]]
     _protocol: Protocol
+    _mapping: ComponentMapping | list[ComponentMapping] | None
+    _name: str | None
 
     def __init__(
         self,
         stateA: ChemicalSystem,
         stateB: ChemicalSystem,
         protocol: Protocol,
-        mapping: Optional[Union[ComponentMapping, list[ComponentMapping], dict[str, ComponentMapping]]] = None,
-        name: Optional[str] = None,
+        mapping: ComponentMapping | list[ComponentMapping] | dict[str, ComponentMapping] | None = None,
+        name: str | None = None,
     ):
-        r"""Two chemical states with a method for estimating free energy difference
+        r"""Two chemical states with a method for estimating the free energy
+        difference between them.
 
-        Connects two :class:`.ChemicalSystem` objects, with directionality,
-        and relates this to a :class:`.Protocol` which will provide an estimate of
-        the free energy difference of moving between these systems.
-        Used as an edge of an :class:`.AlchemicalNetwork`.
+        Connects two :class:`.ChemicalSystem` objects, with directionality, and
+        relates these to a :class:`.Protocol` which will provide an estimate of
+        the free energy difference between these systems. Used as an edge of an
+        :class:`.AlchemicalNetwork`.
 
         Parameters
         ----------
-        stateA, stateB: ChemicalSystem
-           The start (A) and end (B) states of the transformation
-        protocol: Protocol
-           The method used to estimate the free energy difference between states
-           A and B
+        stateA, stateB : ChemicalSystem
+            The start (A) and end (B) states of the transformation.
+        protocol : Protocol
+            The method used to estimate the free energy difference between
+            states A and B.
         mapping : Optional[Union[ComponentMapping, list[ComponentMapping]]]
-           the details of any transformations between :class:`.Component` \s of
-           the two states
+            The details of any transformations between :class:`.Component` \s
+            of the two states.
         name : str, optional
-           a human-readable tag for this transformation
+            A human-readable name for this transformation.
+
         """
         if isinstance(mapping, dict):
             warnings.warn(
-                ("mapping input as a dict is deprecated, " "instead use either a single Mapping or list"),
+                ("mapping input as a dict is deprecated; instead use either a single Mapping or list"),
                 DeprecationWarning,
             )
             mapping = list(mapping.values())
 
         self._stateA = stateA
         self._stateB = stateB
+        self._protocol = protocol
         self._mapping = mapping
         self._name = name
 
-        self._protocol = protocol
-
-    @classmethod
-    def _defaults(cls):
-        return super()._defaults()
-
     def __repr__(self):
-        return f"{self.__class__.__name__}(stateA={self.stateA}, " f"stateB={self.stateB}, protocol={self.protocol})"
+        return f"{self.__class__.__name__}(stateA={self.stateA}, stateB={self.stateB}, protocol={self.protocol}, name={self.name})"
 
     @property
     def stateA(self) -> ChemicalSystem:
@@ -80,31 +211,9 @@ class Transformation(GufeTokenizable):
         return self._stateB
 
     @property
-    def protocol(self) -> Protocol:
-        """The protocol used to perform the transformation.
-
-        This protocol estimates the free energy differences between ``stateA``
-        and ``stateB`` :class:`.ChemicalSystem` objects. It includes all details
-        needed to perform required simulations/calculations and encodes the
-        alchemical pathway used.
-        """
-        return self._protocol
-
-    @property
-    def mapping(self) -> Optional[Union[ComponentMapping, list[ComponentMapping]]]:
+    def mapping(self) -> ComponentMapping | list[ComponentMapping] | None:
         """The mappings relevant for this Transformation"""
         return self._mapping
-
-    @property
-    def name(self) -> Optional[str]:
-        """
-        Optional identifier for the transformation; used as part of its hash.
-
-        Set this to a unique value if adding multiple, otherwise identical
-        transformations to the same :class:`AlchemicalNetwork` to avoid
-        deduplication.
-        """
-        return self._name
 
     def _to_dict(self) -> dict:
         return {
@@ -115,15 +224,11 @@ class Transformation(GufeTokenizable):
             "name": self.name,
         }
 
-    @classmethod
-    def _from_dict(cls, d: dict):
-        return cls(**d)
-
     def create(
         self,
         *,
-        extends: Optional[ProtocolDAGResult] = None,
-        name: Optional[str] = None,
+        extends: ProtocolDAGResult | None = None,
+        name: str | None = None,
     ) -> ProtocolDAG:
         """
         Returns a ``ProtocolDAG`` executing this ``Transformation.protocol``.
@@ -137,101 +242,70 @@ class Transformation(GufeTokenizable):
             transformation_key=self.key,
         )
 
-    def gather(self, protocol_dag_results: Iterable[ProtocolDAGResult]) -> ProtocolResult:
-        """
-        Gather multiple ``ProtocolDAGResult`` into a single ``ProtocolResult``.
 
-        Parameters
-        ----------
-        protocol_dag_results : Iterable[ProtocolDAGResult]
-            The ``ProtocolDAGResult`` objects to assemble aggregate quantities
-            from.
-
-        Returns
-        -------
-        ProtocolResult
-            Aggregated results from many ``ProtocolDAGResult`` objects, all from
-            a given ``Protocol``.
-
-        """
-        return self.protocol.gather(protocol_dag_results=protocol_dag_results)
-
-    def dump(self, file):
-        """Dump this Transformation to a JSON file.
-
-        Note that this is not space-efficient: for example, any
-        ``Component`` which is used in both ``ChemicalSystem`` objects will be
-        represented twice in the JSON output.
-
-        Parameters
-        ----------
-        file : Union[PathLike, FileLike]
-            a pathlike of filelike to save this transformation to.
-        """
-        with ensure_filelike(file, mode="w") as f:
-            json.dump(self.to_dict(), f, cls=JSON_HANDLER.encoder, sort_keys=True)
-
-    @classmethod
-    def load(cls, file):
-        """Create a Transformation from a JSON file.
-
-        Parameters
-        ----------
-        file : Union[PathLike, FileLike]
-            a pathlike or filelike to read this transformation from
-        """
-        with ensure_filelike(file, mode="r") as f:
-            dct = json.load(f, cls=JSON_HANDLER.decoder)
-
-        return cls.from_dict(dct)
-
-
-# we subclass `Transformation` here for typing simplicity
-class NonTransformation(Transformation):
-    """A non-alchemical edge of an alchemical network.
-
-    A "transformation" that performs no transformation at all.
-    Technically a self-loop, or an edge with the same ``ChemicalSystem`` at
-    either end.
-
-    Functionally used for applying a dynamics protocol to a ``ChemicalSystem``
-    that performs no alchemical transformation at all. This allows e.g.
-    equilibrium MD to be performed on a ``ChemicalSystem`` as desired alongside
-    alchemical protocols between it and and other ``ChemicalSystem`` objects.
-    """
+class NonTransformation(TransformationBase):
+    _system: ChemicalSystem
+    _protocol: Protocol
+    _name: str | None
 
     def __init__(
         self,
         system: ChemicalSystem,
         protocol: Protocol,
-        name: Optional[str] = None,
+        name: str | None = None,
     ):
+        """A non-alchemical edge of an alchemical network.
+
+        A "transformation" that performs no transformation at all.
+        Technically a self-loop, or an edge with the same ``ChemicalSystem`` at
+        either end.
+
+        Functionally used for applying a dynamics protocol to a ``ChemicalSystem``
+        that performs no alchemical transformation at all. This allows e.g.
+        equilibrium MD to be performed on a ``ChemicalSystem`` as desired alongside
+        alchemical protocols between it and and other ``ChemicalSystem`` objects.
+
+        Parameters
+        ----------
+        system : ChemicalSystem
+            The system to be sampled, acting as both the starting and end state of the ``NonTransformation``.
+        protocol : Protocol
+            The sampling method to use on the ``system``
+        name : str, optional
+            A human-readable name for this transformation.
+
+        """
 
         self._system = system
-        self._name = name
         self._protocol = protocol
+        self._name = name
+
+    def __repr__(self):
+        return f"{self.__class__.__name__}(system={self.system}, protocol={self.protocol}, name={self.name})"
 
     @property
-    def stateA(self):
+    def stateA(self) -> ChemicalSystem:
+        """The :class:`.ChemicalSystem` this ``NonTransformation`` samples.
+
+        Synonomous with ``system`` attribute and identical to ``stateB``.
+
+        """
         return self._system
 
     @property
-    def stateB(self):
+    def stateB(self) -> ChemicalSystem:
+        """The :class:`.ChemicalSystem` this ``NonTransformation`` samples.
+
+        Synonomous with ``system`` attribute and identical to ``stateA``.
+
+
+        """
         return self._system
 
     @property
     def system(self) -> ChemicalSystem:
+        """The :class:`.ChemicalSystem` this "transformation" samples."""
         return self._system
-
-    @property
-    def protocol(self):
-        """
-        The protocol for sampling dynamics of the `ChemicalSystem`.
-
-        Includes all details needed to perform required
-        simulations/calculations.
-        """
-        return self._protocol
 
     def _to_dict(self) -> dict:
         return {
@@ -240,19 +314,18 @@ class NonTransformation(Transformation):
             "name": self.name,
         }
 
-    @classmethod
-    def _from_dict(cls, d: dict):
-        return cls(**d)
-
     def create(
         self,
         *,
-        extends: Optional[ProtocolDAGResult] = None,
-        name: Optional[str] = None,
+        extends: ProtocolDAGResult | None = None,
+        name: str | None = None,
     ) -> ProtocolDAG:
         """
-        Returns a ``ProtocolDAG`` executing this ``Transformation.protocol``.
+        Returns a ``ProtocolDAG`` executing this ``NonTransformation.protocol``.
         """
+        # TODO: once we have an implicit component mapping concept, use this
+        # here instead of None to allow use of alchemical protocols with
+        # NonTransformations
         return self.protocol.create(
             stateA=self.system,
             stateB=self.system,